feat: Move byDevice page to its own file
