+++ src/app/energyCalculator/byDevice.tsx
@@ -1,0 +1,49 @@
+function CalculatorDeviceRow(){
+  return (
+    <div className="flex flex-row w-full gap-x-4 px-1">
+      {/* Device name */}
+      <div className="flex flex-col w-full gap-y-1">
+        <label htmlFor="devicetype" className="font-semibold">Perangkat</label>
+        <input 
+          type="text" 
+          className="border py-2 px-2"
+          id="deviceType"
+          />
+      </div>
+
+      {/* Device Number */}
+      <div className="flex flex-col w-full gap-y-1">
+        <label htmlFor="deviceNumber" className="font-semibold">Jumlah</label>
+        <input 
+          type="number"
+          id="deviceNumber" 
+          min={0}
+          className="border py-2 px-2"
+        />
+      </div>
+    </div>
+  )
+}
+
+export function CalculatorByDevice(){
+  return (
+    <div className="flex flex-col h-full gap-y-2 w-full">
+      <CalculatorDeviceRow />
+
+      {/* Add button */}
+      <div className="flex justify-center">
+        <button>Tambah Perangkat</button>
+      </div>
+
+      {/* Result */}
+      <div className="flex flex-col w-full px-1 gap-y-2">
+        <label htmlFor="deviceCalculationResult" className="text-center font-semibold">Hasil</label>
+        <input 
+          type="number" 
+          id="deviceCalculationResult" 
+          className="border py-2 bg-gray-400 px-2"
+          disabled />
+      </div>
+    </div>
+  )
+}
